Solar_Access: converted from "facade" pattern to "factory" pattern.
Solar_Access: [BRK] Is now a factory class, not a facade, and returns adapter instances instead of an instance of itself.  Although I note this as a BC-break, you should be able to continue using it almost exactly the way you have; i.e., through Solar::factory().  The only difference should be that you get back an adapter instance instead of the facade instance.

Solar_Access_Adapter: [BRK] Now contains all the methods and properties previously contained in the Facade wrapper.  I note this as a BC break, but if you have been using Solar::factory() all along, you should experience few if any problems on upgrade.

Solar_Access_Adapter_File: [CHG] The fetch() method now checks to see if $this->_config['file'] exists and throws an exception if it does not.