Global removal of func_num_args() and func_get_arg()
In the interest of speeding things up when possible, this change introduces a new global constant: SOLAR_IGNORE_PARAM.  This constant is used when we need an optional method param that may or may not be present.

Previously, we would check for added optional params using func_num_args(), and then get the values of those params using func_get_arg().  The func_*() functions are known to be quite slow.

Now, we set the param default value to SOLAR_IGNORE_PARAM.  If the param is that value, we know it was not passed.

Why do this instead of using null or some other empty value?  Because sometimes you want to pass a null or empty value; using the SOLAR_IGNORE_PARAM value lets you state that normally no parameter is to be passed at all.

Hope this makes sense to everyone.

fixes #11