Solar_Sql_Adapter: [FIX] For PDO in PHP 5.2.1, to better support automated binding of values.
With PDO in PHP 5.2.1 and later, we can no longer just throw an array
of data at the statement; we need to bind values specifically to their
respective placeholders.

In addition, we can't bind one value to multiple identical named
placeholders; we need to bind that same value multiple times. So if
:foo is used three times, PDO uses :foo the first time, :foo2 the
second time, and :foo3 the third time.

This query() method examins the statement for all :name placeholders
and attempts to bind data from the $data array.  The regular-expression
it uses is a little braindead; it cannot tell if the :name placeholder
is literal text or really a place holder.

As such, you should *either* use the $data array for named-placeholder
value binding at query() time, *or* bind-as-you-go when building the 
statement, not both.  If you do, you are on your own to make sure
that nothing looking like a :name placeholder exists as literal text.

Question-mark placeholders are no longer supported for automatic 
value binding at query() time.