Solar_Controller_Page
This fix is to address the "unknown action with custom format" bug discovered
by Nate Abele, and discussed on IRC.

The problem is that we have to pass all the way through to _notFound() before
we can intercept the action. Say the request URI is '/foo/bar.xml', and the
Vendor_App_Foo class has no actionBar() method. Currently, the flow will pass
to _notFound(), which can then forward to some other method (e.g.,
actionIndex() or something like that). Although the action we forwarded to is
allowed an XML format, the original action (bar) was not allowed an XML
format, so $_format was dropped and replaced with the default format.

This change adds a _fixAction() method as part of the _load() process, so you
can intercept the action and change it before the format is determined. By
default, it is empty, so the old behavior remains; however, you can override
it to make it do whatever you like.

* [CHG] Move "find the requested action" logic from _loadAction() into
  _loadInfoQueryFormat().

* [DEL] Remove the _loadAction() method entirely.

* [ADD] Empty method _fixAction() as part of the _load() process to allow
  extended classes to intercept and change the requested action earlier.