Solar_Sql_Adapter_MysqlReplicated: BIG FIX, READ CAREFULLY
Until this change, the adapter **always** connected to the slave, **never** the master, because of a copy & paste error in the _setDsn() method.  Previously, the master used $_dsn, not $_dsn_master, for its connection information.  This change fixes that.  One reason I didn't notice it before now is that MySQL slaves accept writes by default, instead of being read-only by default, but that is a secondary fault -- the primary fault is still mine.

* [FIX] In method _setDsn(), use $_dsn_master, not $_dsn, for the master PDO object.

* [CHG] Retain connection information the same way as the base adapter.