Solar_Model_Tags_Collection: [FIX] Method setNames() now correctly removes all tags, not just every-other one.
This is where SPL iterators (at least the way we're using them) become trouble.  If you are iterating over $this internally, it will interact with *external* iterators because iteration properties are all on the same instance.  Thus, to iterate over $this internally, you have to clone($this) and use the clone for iteration, while performing any necessary actions on $this.  Confusing and unintuitive, until you remember how SPL iterators work.  Perhaps there is some better implementation that I'm currently unaware of.